#328: Moved the equipment models to model/items/equipment
-
