+++ src/app/ffbe/mappers/equipment-mapper.spec.ts
@@ -1,8 +1,8 @@
 import {Objet} from '../model/objet/objet.model';
 import {FfbeUtils} from '../utils/ffbe-utils';
-import {EQUIPMENTS_TEST_DATA} from '../model/equipment/equipment.model.spec';
+import {EQUIPMENTS_TEST_DATA} from '../model/items/equipment/equipment.model.spec';
 import {EquipmentMapper} from './equipment-mapper';
-import {Equipment} from '../model/equipment/equipment.model';
+import {Equipment} from '../model/items/equipment/equipment.model';
 import {CharactersServiceMock} from '../services/characters.service.spec';
 import {CharactersService} from '../services/characters.service';
 import {CHARACTER_TEST_DATA} from '../model/character/character.model.spec';
