+++ src/app/ffbe/mappers/skill-mapper.ts
@@ -3,7 +3,7 @@ import {Skill} from '../model/skill.model';
 import {Competence} from '../model/competence.model';
 import {HTML_LINE_RETURN, SkillEffectsMapper} from './effects/skill-effects.mapper';
 import {EquipmentsService} from '../services/equipments.service';
-import {Equipment} from '../model/equipment/equipment.model';
+import {Equipment} from '../model/items/equipment/equipment.model';
 import {MateriasService} from '../services/materias.service';
 import {Materia} from '../model/materia.model';
 import {FfbeUtils} from '../utils/ffbe-utils';
