+++ src/app/ffbe/model/items/equipment/equipment-element-resist.model.ts
