+++ src/app/ffbe/model/items/equipment/equipment-stats.model.ts
