+++ src/app/ffbe/model/items/equipment/equipment-status-effect.model.ts
