+++ src/app/ffbe/model/items/equipment/equipment-strings.model.ts
