+++ src/app/ffbe/model/items/equipment/equipment.model.spec.ts
