+++ src/app/ffbe/services/equipments.service.spec.ts
@@ -1,4 +1,4 @@
-import {Equipment} from '../model/equipment/equipment.model';
+import {Equipment} from '../model/items/equipment/equipment.model';
 
 export class EquipmentsServiceMock {
   private static INSTANCE: EquipmentsServiceMock = new EquipmentsServiceMock();
